Storing/restoring parsed data context via serialization.
When exporting several related(but unmerged) bzr branches into a git
repository, we:

* Export 'trunk' bzr branch into a git 'master', creating bzr and
  git marks files. This operation involves exporting all trunk revisions
  and is time consuming.

* Export all bzr branches (having 'trunk' as their parent)
  into corresponding git branches. Previously generated mark files are
  used in order to import only branch-specific revisions. This operation
  is usually faster, because only branch-with-trunk difference is
  imported.

In order to process (2), Bzr2Git4Notes needs to build full commit tree,
including data received on (1). So, instead of doing each time
time-consuming re-parsing 'trunk' commit history, the utility now
serializes it on (1) and de-serializes on (2).