Added helpers/merged_branches.sh.
While importing non-official branches, it is possible that some of these
branches were fully merged into another branches.  That means that other
branches contain all revisions a 'merged' branch has. As a result,
during import, these branches can be skipped (not created) in
git.

This script compiles a list of such 'merged' bzr branches.
Having this list, it is possible to:
- either import them first and have them in the result git repository,
- or import them last, thus skipping and considering as 'unnecessary'.